+++ doc/references/release-notes.rst
@@ -17,6 +17,12 @@ Bug Fixes
 * Fixed issue when copying a network with an unsolved model.
   (https://github.com/PyPSA/PyPSA/pull/5)
 
+* Fixed missing dependency issue for `typing-extensions`
+  (https://github.com/PyPSA/PyPSA/pull/1264)
+
+* Fixed Excel import when snapshots sheet is missing 
+  (https://github.com/PyPSA/PyPSA/issues/1268)
+
 `v0.35.0 <https://github.com/PyPSA/PyPSA/releases/tag/v0.35.0>`__ (22th June 2025)
 =======================================================================================